requirements.txt
requests~=2.32.3
functions_framework~=3.8.2
flask~=3.1.0
google-cloud-secret-manager